# backend/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Bus Reservation System API
  description: |
    Comprehensive API documentation for the Bus Reservation System. 
    Includes:
    1. Public Search API (Searching buses, filtering, sorting).
    2. Operator API (Manifests, bookings, passenger lists).
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local Development Server
paths:
...
  /operator/buses/{bus_id}/seats:
    put:
      summary: Update Seat Layout (Operator Only)
      description: |
        Allows the Operator to update price, type, or status for specific seats in their bus.
        Useful for defining the bus layout (e.g., setting the first 10 seats as Sleeper).
      operationId: updateBusLayout
      tags:
        - Operator
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bus_id
          schema:
            type: integer
          required: true
          description: The ID of the bus to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSeatsRequest'
      responses:
        '200':
          description: Seats updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateSeatsResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '403':
          description: Forbidden - User is not an Operator or does not own this bus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Bus not found
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      summary: Get Bus Layout (Operator Only)
      description: Fetches all seats for a specific bus. Used to populate the "Edit Layout" dashboard.
      operationId: getBusLayout
      tags:
        - Operator
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bus_id
          schema:
            type: integer
          required: true
          description: ID of the bus to fetch
      responses:
        '200':
          description: List of seats
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        seat_id:
                          type: integer
                        seat_number:
                          type: string
                        seat_type:
                          type: string
                        price:
                          type: number
                        status:
                          type: string
        '403':
          description: Forbidden - You do not own this bus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SeatUpdateItem:
      type: object
      required:
        - seat_numbers
      properties:
        seat_numbers:
          type: array
          description: List of seat numbers to apply the changes to
          items:
            type: string
          example: ["S1", "S2", "S3", "S4"]
        price:
          type: number
          format: float
          description: New price for this seat
          example: 550.00
        seat_type:
          type: string
          enum: ['Seater', 'Sleeper_Upper', 'Sleeper_Lower']
          example: "Sleeper_Lower"
        status:
          type: string
          enum: ['good', 'repair', 'maintenance', 'damaged']
          example: "good"
    UpdateSeatsRequest:
      type: object
      required:
        - seats
      properties:
        updates:
          type: array
          description: List of seat updates to apply
          items:
            $ref: '#/components/schemas/SeatUpdateItem'
    UpdateSeatsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Successfully updated 10 seats"

    # --- Route Management Schemas ---
...
    ErrorResponse:
      type: object
      description: Standard error wrapper for all non-2xx responses.
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid date format."
        code:
          type: string
          example: "VAL_001"
      required: [status, error, message]
  responses:
    BadRequestError:
      description: Bad Request – Invalid inputs (e.g., date format or missing ID).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error – unexpected condition on the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'